requirements.txt
pandas
scikit-learn
google-cloud-aiplatform
joblib